fix: zero warnings, macOS crash handler, smoke test
- Fix all 155 compiler warnings (95 unique sources) across 49 files:
  unused params, missing override, initializer reorder, sign compare,
  non-virtual destructors, deprecated APIs, operator precedence bugs
- Fix real bug in cell_entity.cpp:292 where operator precedence made
  the early-return dead code (& binds tighter than |)
- Implement signal-based crash handler for non-Windows (SIGSEGV,
  SIGBUS, SIGABRT, SIGFPE) using sigaction + backtrace
- Add smoke-test script and `make smoke-test` target to validate
  binary loads, reads config, connects to DB, opens listening port
- Suppress warnings on third-party Recast/Detour FetchContent targets

+++ src/nav_builder/mesh.cpp
@@ -142,7 +142,7 @@ unsigned int Mesh::translateFaceIndex(std::string const & index)
 	else
 		val = atol(index.c_str());
 
-	if (val == 0 || (val > 0 && val > vertices_.size()) || (val < 0 && -val > vertices_.size()))
+	if (val == 0 || (val > 0 && static_cast<size_t>(val) > vertices_.size()) || (val < 0 && static_cast<size_t>(-val) > vertices_.size()))
 		throw std::runtime_error("Invalid vertex index in object!");
 
 	if (val < 0)
